pyproject.toml
[tool.poetry]
name = "obs-control"
version = "0.1.0"
description = ""
authors = ["Serhii Khalymon <sergiykhalimon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
fbs = {git = "https://github.com/Infernion/fbs"}
pyqt5 = "5.9.2"
macholib = "^1.11"
sip = "4.19.8"
obs-websocket-py = "^0.4"
enaml = "^0.10.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
